Fix issue with parsing date with hardcoded format
We are creating a DateTimeOffset on the server and sending it in string format to the client. There, using ParseExact and a hardcoded US date format, we are converting it back to a DateTimeOffset.

The problem is that when the date is created on a server that is in a non-US locale, the client throws and error and the app crashes.